mata: Allow disabling camera laser sensor in developer options
Sometimes, being able to disable the camera laser sensor quickly is
useful for taking pictures of reflective objects without a purple dot
showing up in the picture, especially when precise upfront focus isn't
required. Enable the existing toggle in developer options to expose this
functionality.

Older versions of the stock ROM included this option, but it's been
removed since then.

Change-Id: I24fcde77536fffcdafda749bd02157c510ac9aa1